# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LISAN MVP API
  version: 1.1.0
  description: Auth, Placement, Learning Path, Gamification, Subscriptions, and AI tutor MVP contracts.
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /analytics/events:
    post:
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventName]
              properties:
                eventName:
                  type: string
                  enum:
                    - ai_tutor_submitted
                    - ai_tutor_success
                    - ai_tutor_retry
                    - ai_tutor_cooldown_hit
                    - ai_tutor_daily_cap_hit
                source: { type: string, default: web }
                metadata: { type: object }
      responses:
        '201': { description: Event stored }
  /analytics/summary:
    get:
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: hours
          schema: { type: integer, minimum: 1, maximum: 168, default: 24 }
        - in: query
          name: scenario
          schema: { type: string, enum: [all, daily, travel, work, migration], default: all }
      responses:
        '200': { description: Event counts by type for selected time window }
  /analytics/trend:
    get:
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: hours
          schema: { type: integer, minimum: 1, maximum: 168, default: 24 }
        - in: query
          name: scenario
          schema: { type: string, enum: [all, daily, travel, work, migration], default: all }
      responses:
        '200': { description: Hourly trend points for analytics events }
  /analytics/ops/dashboard:
    get:
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: hours
          schema: { type: integer, minimum: 1, maximum: 168, default: 24 }
      responses:
        '200':
          description: Operational KPIs and hourly trend (latency, errors, retries, circuit state)
  /analytics/ops/routes:
    get:
      tags: [Analytics]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: hours
          schema: { type: integer, minimum: 1, maximum: 168, default: 24 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 5, maximum: 100, default: 20 }
      responses:
        '200':
          description: Route-level SLO snapshot (P95 latency + 5xx by endpoint + top failing routes)